update PersonalId and PhoneNumber to type long.

+++ Egzaminas_ZmogausRegistravimoSistema/Entities/PersonInfo.cs
@@ -9,8 +9,8 @@ namespace Egzaminas_ZmogausRegistravimoSistema.Entities
         public long Id { get; set; }
         public string FirstName { get; set; } = string.Empty;
         public string LastName { get; set; } = string.Empty;
-        public int PersonalId { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PersonalId { get; set; }
+        public long PhoneNumber { get; set; }
         public string Email { get; set; } = string.Empty;
         public string PhotoPath { get; set; } = string.Empty;
